# Rewrite the GPU Lightmapper's indirect logic to match Godot 3.5's CPU Lightmapper.