--- a/modules/lightmapper_rd/lm_common_inc.glsl
+++ b/modules/lightmapper_rd/lm_common_inc.glsl
@@ -1,6 +1,26 @@
 
 /* SET 0, static data that does not change between any call */
 
+layout(set = 0, binding = 0) uniform BakeParameters {
+	vec3 world_size;
+	float bias;
+
+	vec3 to_cell_offset;
+	int grid_size;
+
+	vec3 to_cell_size;
+	uint light_count;
+
+	mat3x4 env_transform;
+
+	ivec2 atlas_size;
+	float exposure_normalization;
+	uint bounces;
+
+	float bounce_indirect_energy;
+}
+bake_params;
+
 struct Vertex {
 	vec3 position;
 	float normal_z;
@@ -51,9 +71,10 @@ struct Light {
 	float cos_spot_angle;
 	float inv_spot_attenuation;
 
+	float indirect_energy;
 	float shadow_blur;
 	bool static_bake;
-	uint pad[2];
+	uint pad;
 };
 
 layout(set = 0, binding = 4, std430) restrict readonly buffer Lights {