--- a/modules/lightmapper_rd/register_types.cpp
+++ b/modules/lightmapper_rd/register_types.cpp
@@ -46,10 +46,10 @@ void initialize_lightmapper_rd_module(ModuleInitializationLevel p_level) {
 		return;
 	}
 
-	GLOBAL_DEF("rendering/lightmapping/bake_quality/low_quality_ray_count", 16);
-	GLOBAL_DEF("rendering/lightmapping/bake_quality/medium_quality_ray_count", 64);
-	GLOBAL_DEF("rendering/lightmapping/bake_quality/high_quality_ray_count", 256);
-	GLOBAL_DEF("rendering/lightmapping/bake_quality/ultra_quality_ray_count", 1024);
+	GLOBAL_DEF("rendering/lightmapping/bake_quality/low_quality_ray_count", 32);
+	GLOBAL_DEF("rendering/lightmapping/bake_quality/medium_quality_ray_count", 128);
+	GLOBAL_DEF("rendering/lightmapping/bake_quality/high_quality_ray_count", 512);
+	GLOBAL_DEF("rendering/lightmapping/bake_quality/ultra_quality_ray_count", 2048);
 	GLOBAL_DEF("rendering/lightmapping/bake_performance/max_rays_per_pass", 32);
 	GLOBAL_DEF("rendering/lightmapping/bake_performance/region_size", 512);