… modules/lightmapper_rd/lightmapper_rd.cpp
… modules/lightmapper_rd/lm_compute.glsl