## Problem

The current lightmapper suffers from some implementation flaws that are very tough to fix without rewriting the logic.

One issue that has been repeatedly brought up is https://github.com/godotengine/godot/issues/73295, which clearly shows the wrong behavior of the indirect lighting logic as bounces increases in the scene. My conclusions of why the wrong behavior was caused [were pretty apparent when digging into the implementation](https://github.com/godotengine/godot/issues/73295#issuecomment-1712000939).

For anyone unfamiliar with the topic, the expectation when increasing bounces is not for the overall ambience of the scene to increase beyond the first few iterations. It is very rare for any lighting solutions out there to achieve anything significant past six bounces or so due to how the light transfer works on each step.

Godot 3.x's CPU lightmapper [has a much more correct implementation of how this should behave](https://github.com/godotengine/godot/blob/fc32e066af1cd6766762dec31c7d2224f3d42c5f/modules/lightmapper_cpu/lightmapper_cpu.cpp#L840). It's worth noting that while it's not exactly physically correct, it has the nice benefit of providing a user-tweakable attribute to get as much indirect energy in the bounces as the user wants, which this PR also adds as a configurable option (thanks to @Calinou for the reference PR at https://github.com/godotengine/godot/pull/51705).

As I did not feel it'd be right to rewrite the current implementation with one of my own in this particular case, it seemed to be a better fit to just port the behavior from Godot 3.x so it works pretty much the same way in 4.x.

## Solution

This PR reimplements the logic used by both the indirect bounces step and the light probes step to match the behavior of Godot 3.x, but does it in the existing compute shader. There's a noticeable reduction in baking times, as extra bounces are not calculated by just iterating on top of the starting position and can be terminated early if it hits the environment or a back-face. It also implements Godot 3.5's Russian Roulette early termination for an additional optimization based on the probability of the current color being carried through the bounce.

## Comparisons

All shots are taken with the same settings (high quality) except for the bounce. I would've gone higher but current master branch was taking far too long to bake it due to each bounce counting as a full iteration. Denoiser is off.

Given the nature of the algorithm has changed, the **noise will be decreased further and further in master compared to this PR** because it just traces more rays (as evident by the longer bake times). **Just because master looks less noisy under the same quality setting doesn't imply the result is better**, as it lacks the ability to converge to the correct result.

This behavior is expected and can be easily compensated by just increasing the ray count in the PR instead, as each bounce no longer implies another set of rays being dispatched but just an increase of cost per ray instead.

### PR - 1 Bounce

Time to bake: 01:22

![PR_BOUNCES_1](https://github.com/godotengine/godot/assets/538504/d3de423c-08f0-4289-9514-c96d18a03204)

### PR - 3 Bounces

The biggest benefits to indirect lighting are indeed achieved with at least one more additional bounce, as evident here.

Time to bake: 01:40

![PR_BOUNCES_3](https://github.com/godotengine/godot/assets/538504/9a95e53b-b2f2-4cfc-9dc5-c909e2600563)

### PR - 6 Bounces

As expected, it hardly looks any different from the 3 bounces version.

Time to bake: 01:51

![PR_BOUNCES_6](https://github.com/godotengine/godot/assets/538504/98251cae-c6ad-4e82-ba84-dd289c566061)

As a bonus to showcase the noise difference I mentioned, here's an Ultra quality shot of the PR, showing the convergence.

### Bonus Shot - PR - 6 Bounces - Ultra Quality

Just to showcase the difference that can be achieved with increasing the ray count and how it converges to a much better result.

![PR_BOUNCES_6_ULTRA](https://github.com/godotengine/godot/assets/538504/8f4867f9-e56f-4bb7-900a-ea8749d82a96)

### Master - 1 Bounce

Expected to be pretty much identical to this PR.

Time to bake: 01:22

![MASTER_BOUNCES_1](https://github.com/godotengine/godot/assets/538504/ca6b813d-e68f-4325-b658-bff3488f2dee)

### Master - 3 Bounces

Lighting starts to get significantly washed out.

Time to bake: 02:48

![MASTER_BOUNCES_3](https://github.com/godotengine/godot/assets/538504/9e06a719-b5a8-4e0e-a914-e3f9da14411d)

### Master - 6 Bounces

We're going full speed ahead towards full bright at this point.

Time to bake: 04:31

![MASTER_BOUNCES_6](https://github.com/godotengine/godot/assets/538504/f9ac2dfd-b273-4a6f-8b97-38aab777c82a)

## Feedback

We should make sure this doesn't break any of the existing bake options, as there's different configurations (static, dynamic, directional, etc) to try out.

Sadly, it is expected the visual result will change when using existing settings and that artists will need to tweak them to get a similar result. There's no realistic way to automatically convert this and it's probably best to just address this as soon as possible given how many other issues this can fix.

However, it's worth noting Lightmap GI will have the new "Bounce Indirect Energy" property, as well as all types of light will actually use the "Indirect Energy" property when baking, which should allow much better control over the intensity of the indirect lighting from the baked result.

## TODO

- [x] Using physical light units seem to be broken.
- [x] Review if directional baking works properly.
- [x] Change ray counts of the quality presets to resemble the approach master takes for the default bounces.
- [x] Change max property hint for bounces to 6.

_Bugsquad edit: Fixes https://github.com/godotengine/godot/issues/77356; Fixes https://github.com/godotengine/godot/issues/73295; Fixes https://github.com/godotengine/godot/issues/77169_