@jcostello 2023-09-21
Here is my test in my lightmap scene

https://imgsli.com/MjA3ODY1
https://imgsli.com/MjA3ODY2
https://imgsli.com/MjA3ODY3
https://imgsli.com/MjA3ODY4
https://imgsli.com/MjA3ODY5
https://imgsli.com/MjA3ODcw

This last 3 are energy comparison in the same PR
Fe de erratas: this one is 3 bounces vs 3 bounces
https://imgsli.com/MjA3ODcy
https://imgsli.com/MjA3ODcz
https://imgsli.com/MjA3ODc1

Seems that the bounces works better but still introduces a lot of light after each bounce

in the description of the PR 3 vs 6 still introduce a lot of sky light and ambient light but IDK the internals of the algoritm. In blender seems that after 4 bounces, only small details get more light.
This is only my opition but in general seems that the lightmap indirect lighting works much much better

Also, probes seems to work better with static lighting. May be fix this issue https://github.com/godotengine/godot/issues/77356