@jcostello 2023-09-21
@DarioSamo yeah, surface is pure white. I imagine that the length of the ray has to be taken into consideration for the atenuation after the bounce, even if its pure white.

Never the less the PR fixes most of the cases and probably #77356 (I have to test it further)

There is this case we have to test and figure it out what is going on before merge

![Screenshot from 2023-09-21 14-28-46](https://github.com/godotengine/godot/assets/1168582/ee8a59a9-00f5-46fd-91cc-5cf5782b61aa)