@jcostello 2023-09-21
Static
![Screenshot from 2023-09-21 16-54-05](https://github.com/godotengine/godot/assets/1168582/f5993210-5395-496a-9eb6-193c788de321)

Dynamic
![Screenshot from 2023-09-21 16-55-07](https://github.com/godotengine/godot/assets/1168582/2d618575-a622-42e8-9611-a3f489b66459)

Seems to fix https://github.com/godotengine/godot/issues/77356 for probes generation on static vs dynamic. Still need to fix invalid probes