@Calinou 2023-09-22
## Bake times comparison

**OS:** Fedora 38
**CPU:** Intel Core i9-13900K
**GPU:** GeForce RTX 4090

Using an optimized editor build (`optimize=speed lto=full`). Project uses 3 bounces in LightmapGI, and has Generate Probes subdiv set to 8 (the default). Project settings were left at their defaults.

I can confirm light probes now have the correct brightness level with this PR.

**Testing project:** [test_dm3_open_sky.zip](https://github.com/godotengine/godot/files/12695838/test_dm3_open_sky.zip)

Quality | Before | After *(this PR)*
-|-|-
High | 9 seconds | 9 seconds
Ultra | 25 seconds | 25 seconds

## Visual comparison

### High

Before | After *(this PR)*
-|-
![Screenshot_20230922_033615_high_before_1 webp](https://github.com/godotengine/godot/assets/180032/a94de852-d7b5-4eb5-9493-58e965944e56) | ![Screenshot_20230922_033846_high_after_1 webp](https://github.com/godotengine/godot/assets/180032/2ef6ff1c-03b3-4b75-ac69-e6c5b5e9117b)

Before | After *(this PR)*
-|-
![Screenshot_20230922_033625_high_before_2 webp](https://github.com/godotengine/godot/assets/180032/ded062d0-7a09-4a2c-a435-8a22130752ac) | ![Screenshot_20230922_033849_high_after_2 webp](https://github.com/godotengine/godot/assets/180032/f90db283-e681-41c7-b70c-a914bd33d003)

Before | After *(this PR)*
-|-
![Screenshot_20230922_033629_high_before_3 webp](https://github.com/godotengine/godot/assets/180032/4e530287-1e3c-41ac-8bee-9187a1012058) | ![Screenshot_20230922_033901_high_after_3 webp](https://github.com/godotengine/godot/assets/180032/953fccfa-9a37-42e5-9c69-f0316b9664b1)

Before | After *(this PR)*
-|-
![Screenshot_20230922_033634_high_before_4 webp](https://github.com/godotengine/godot/assets/180032/82d4757f-4fb6-4da1-830c-b906d7993484) | ![Screenshot_20230922_033906_high_after_4 webp](https://github.com/godotengine/godot/assets/180032/db0fd541-8ac9-40a3-84a7-bc31ccf75f62)

### Ultra

Before | After *(this PR)*
-|-
![Screenshot_20230922_033724_ultra_before_1 webp](https://github.com/godotengine/godot/assets/180032/49853ac0-affc-4458-a0f4-e3da778dffdf) | ![Screenshot_20230922_034010_ultra_after_1 webp](https://github.com/godotengine/godot/assets/180032/6f14da30-c49d-487e-bca2-957c88cdf761)

Before | After *(this PR)*
-|-
![Screenshot_20230922_033727_ultra_before_2 webp](https://github.com/godotengine/godot/assets/180032/aa761f77-a6de-4371-813c-6dd6d0f39c2b) | ![Screenshot_20230922_034013_ultra_after_2 webp](https://github.com/godotengine/godot/assets/180032/6e263c77-903f-4cb4-bf3c-37f73340c721)

Before | After *(this PR)*
-|-
![Screenshot_20230922_033731_ultra_before_3 webp](https://github.com/godotengine/godot/assets/180032/f1d7c9ed-416a-414e-8a09-1eca1f2e187e) | ![Screenshot_20230922_034016_ultra_after_3 webp](https://github.com/godotengine/godot/assets/180032/518010f4-b685-4454-afeb-41402676ca73)

Before | After *(this PR)*
-|-
![Screenshot_20230922_033735_ultra_before_4 webp](https://github.com/godotengine/godot/assets/180032/c10ec7b0-aeca-4aea-bb0e-36d4af1cf581) | ![Screenshot_20230922_034020_ultra_after_4 webp](https://github.com/godotengine/godot/assets/180032/9e37895e-7e0a-4974-a0d8-f7cb31a848b9)